+++ SpiderCam/optoWheel.h
@@ -20,7 +20,7 @@ public:
   virtual ~optoWheel();
   void reset(void);
   void pinAssign(int encA, int encB);
-  int get_posAbsPasos(void);
+  long get_posAbsPasos(void);
   int get_direccionAct(void);
   int get_dientesPorVuelta(void);
 //  float get_mmPorVuelta(void)
